Redo limits to allow the pick and place gripper demo to work again
Looks like if move group commander is used on the gripper, the values need
to be doubled or halved, right now it will always think it has failed though
in reality the gripper will reach the intended pose.

This also might cause trouble down the line, as the joint limits are double
what the gripper can reach right now.  Need to look into the commander further
to see why this is happening